Separado ast y parser
