+++ src/retriever/parser.py
@@ -1,5 +1,4 @@
-from abc import ABC, abstractmethod
-
+from .ast import AndNode, NotNode, OrNode, WordNode
 from .lexer import (
     WORD,
     AndToken,
@@ -10,61 +9,6 @@ from .lexer import (
     OrToken,
     RParenToken,
 )
-
-
-class Node(ABC):
-    @abstractmethod
-    def eval(self, index):
-        ...
-
-
-class AndNode(Node):
-    def __init__(self, left, right):
-        self.left = left
-        self.right = right
-
-    def eval(self, index):
-        return list(set(self.left.eval(index)) & set(self.right.eval(index)))
-
-    def __str__(self):
-        return f"({self.left} AND {self.right})"
-
-
-class OrNode(Node):
-    def __init__(self, left, right):
-        self.left = left
-        self.right = right
-
-    def eval(self, index):
-        return list(
-            sorted(set(self.left.eval(index)) | set(self.right.eval(index)))
-        )
-
-    def __str__(self):
-        return f"({self.left} OR {self.right})"
-
-
-class NotNode(Node):
-    def __init__(self, data):
-        self.data = data
-
-    def eval(self, index):
-        all_docs: set = set(index.ids_all_docs)
-        return list(all_docs - set(self.data.eval(index)))
-
-    def __str__(self):
-        return f"NOT ({self.data})"
-
-
-class WordNode(Node):
-    def __init__(self, data):
-        self.data = data
-
-    def eval(self, index):
-        return index.postings[self.data]
-
-    def __str__(self):
-        return self.data
 
 
 class InvalidQueryException(Exception):
